// src/app/shared/interfaces/mat-columns.ts
export interface MatColumns {
    //@var title String title Titre de la colonne dans le tableau
    title:String;
    //@var always Boolean Determine si la colonne doit être affichée
    always: Boolean;
    //@var value String Valeur utilisée dans la liste des colonnes a afficher
    value: String;
    //@var isDisplayed boolean Vrai si la colonne est affichée 
    isDisplayed: Boolean;
}